refactor plugins file names
