+++ src/plugins/plugineditor.hpp
