+++ src/plugins/pluginprocessor.hpp
